requirements.txt
pandas>=1.5.3
numpy>=1.21.0
scikit-learn>=1.0.2
matplotlib>=3.5.0
seaborn>=0.11.0
kaggle>=1.5.12
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
python-dotenv>=0.19.0
apache-airflow==2.7.1
notebook==7.0.3